\encoding{latin1}
\name{spams.calcXY}
\alias{spams.calcXY}
\title{
spams.calcXY
}
\description{
    Compute Z=XY using the BLAS library used by SPAMS.
}
\usage{
spams.calcXY(X,Y)
}
\arguments{
\item{X}{%
double m x n matrix
\preformatted{%
}}
\item{Y}{%
double n x p matrix   
\preformatted{%
}}
}
\value{
\item{Z}{%
double m x p matrix 
\preformatted{%
}}
}
\author{
Julien MAIRAL, 2009 (spams, matlab interface and documentation)
Jean-Paul CHIEZE 2011-2012 (R interface)
}
