\encoding{latin1}
\name{spams.calcXYt}
\alias{spams.calcXYt}
\title{
spams.calcXYt
}
\description{
    Compute Z=XY' using the BLAS library used by SPAMS.
}
\usage{
spams.calcXYt(X,Y)
}
\arguments{
\item{X}{%
double m x n matrix
\preformatted{%
}}
\item{Y}{%
double p x n matrix   
\preformatted{%
}}
}
\value{
\item{Z}{%
double m x p matrix 
\preformatted{%
}}
}
\author{
Julien MAIRAL, 2009 (spams, matlab interface and documentation)
Jean-Paul CHIEZE 2011-2012 (R interface)
}
