\encoding{latin1}
\name{spams.decompSimplex}
\alias{spams.decompSimplex}
\title{
spams.decompSimplex
}
\description{
    documentation to appear soon
}
\usage{
spams.decompSimplex(X,Z,computeXtX = FALSE,numThreads = -1)
}
\arguments{
\item{X}{%
double m x n matrix   (input signals)
\preformatted{%
m is the signal size
n is the number of signals to decompose
}}
}
\value{
\item{Z}{%
double %
\preformatted{%
}}
}
\author{
Yuansi Chen and Julien MAIRAL, 2014 (spams, matlab interface and documentation)
Jean-Paul CHIEZE 2011-2012 (R interface)
}
