\encoding{latin1}
\name{spams.invSym}
\alias{spams.invSym}
\title{
spams.invSym
}
\description{
    returns the inverse of a symmetric matrix A
}
\usage{
spams.invSym(A)
}
\arguments{
\item{A}{%
double n x n matrix   
\preformatted{%
}}
}
\value{
\item{B}{%
double n x n matrix 
\preformatted{%
}}
}
\author{
Julien MAIRAL, 2009 (spams, matlab interface and documentation)
Jean-Paul CHIEZE 2011-2012 (R interface)
}
