\encoding{latin1}
\name{spams.normalize}
\alias{spams.normalize}
\title{
spams.normalize
}
\description{
    rescale the columns of X so that they have
       unit l2-norm.
}
\usage{
spams.normalize(X)
}
\arguments{
\item{X}{%
double m x n matrix   
\preformatted{%
}}
}
\value{
\item{Y}{%
double m x n matrix 
\preformatted{%
}}
}
\author{
Julien MAIRAL, 2010 (spams, matlab interface and documentation)
Jean-Paul CHIEZE 2011-2012 (R interface)
}
