\encoding{latin1}
\name{spams.sort}
\alias{spams.sort}
\title{
spams.sort
}
\description{
    sort the elements of X using quicksort
}
\usage{
spams.sort(X,mode=T)
}
\arguments{
\item{X}{%
double vector of size n
\preformatted{%
}}
\item{mode}{%
false for decreasing order (true by default)
\preformatted{%
}}
}
\value{
\item{Y}{%
double  vector of size n
\preformatted{%
}}
}
\author{
Julien MAIRAL, 2010 (spams, matlab interface and documentation)
Jean-Paul CHIEZE 2011-2012 (R interface)
}
